chore: promote v5 templates
