+++ templates/rspack.config.cjs
