+++ templates/rspack.config.mjs
